+++ KiddyEatsApp/Presentation/View/Explore/Components/RecipeCard.swift
@@ -51,7 +51,7 @@ struct RecipeCard: View {
                     
                     Spacer()
                     
-                    SaveToCollectionsButton(babyMeal: babyMeal)
+                    RecipeCardSaveToCollectionsButton(babyMeal: babyMeal)
                         .buttonStyle(KiddyEatsMiniProminentButtonStyle())
                 }
                 .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .init(horizontal: .center, vertical: .top))
